Change:個々のファイルが大きくならないように、一部Widgetを別ファイルに移動
