+++ lib/page/github_repo_detail_page.dart
@@ -1,10 +1,9 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:yumemi_codecheck_repo_search/common/brightness_adaptive_svg.dart';
+import 'package:yumemi_codecheck_repo_search/common/launch_url.dart';
 import 'package:yumemi_codecheck_repo_search/common/loading_indicator.dart';
 import 'package:yumemi_codecheck_repo_search/const.dart';
-import 'package:yumemi_codecheck_repo_search/generated/l10n.dart';
 import 'package:yumemi_codecheck_repo_search/model/owner.dart';
 import 'package:yumemi_codecheck_repo_search/model/repo.dart';
 
@@ -106,7 +105,7 @@ class _RepoDescription extends StatelessWidget {
             ),
             const SizedBox(width: 8),
             IconButton(
-              onPressed: () => _launchUrl(context, repo.htmlUrl),
+              onPressed: () => launchUrlSafe(context, repo.htmlUrl),
               icon: const Icon(Icons.launch),
             ),
           ],
@@ -158,7 +157,7 @@ class _AvatarImage extends StatelessWidget {
   Widget build(BuildContext context) {
     final borderRadius = BorderRadius.circular(32);
     return InkWell(
-      onTap: () => _launchUrl(context, owner.htmlUrl),
+      onTap: () => launchUrlSafe(context, owner.htmlUrl),
       borderRadius: borderRadius,
       child: ClipRRect(
         borderRadius: borderRadius,
@@ -190,19 +189,6 @@ class _SVGAndText extends StatelessWidget {
         const SizedBox(width: 4),
         Text(text),
       ],
-    );
-  }
-}
-
-Future<void> _launchUrl(BuildContext context, String url) async {
-  final uri = Uri.parse(url);
-  try {
-    await launchUrl(uri);
-  } catch (e) {
-    if (!context.mounted) return;
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text(S.current.failedLaunch(url))),
     );
   }
 }
