+++ lib/page/widget/suggestion_view.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:yumemi_codecheck_repo_search/service.dart';
+
+class SuggestionsView extends ConsumerWidget {
+  const SuggestionsView({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return ref.watch(queryHistoryStreamProvider).when(
+          data: (data) {
+            if (data.isEmpty) return const SizedBox();
+
+            return SingleChildScrollView(
+              child: Column(
+                children: [
+                  for (final query in data)
+                    ListTile(
+                      title: Text(query),
+                      leading: const Icon(Icons.history),
+                      trailing: GestureDetector(
+                        onTap: () =>
+                            ref.read(queryHistoryServiceProvider).remove(query),
+                        child: const Icon(Icons.delete),
+                      ),
+                      onTap: () => ref
+                          .read(repoSearchQueryProvider.notifier)
+                          .update(query),
+                    ),
+                ],
+              ),
+            );
+          },
+
+          error: (error, stackTrace) => Text(
+            error.toString(),
+            style: TextStyle(
+              color: Theme.of(context).colorScheme.error,
+            ),
+          ), // TODO: 適切なエラーメッセージを表示,
+          loading: () => const SizedBox(),
+        );
+  }
+}
